extract Strategy class into its own file and rename it to Shaker
